src/main/java/utils/Utils.java
package utils;

import java.io.FileInputStream;
import java.io.IOException;
import java.net.InetSocketAddress;
import java.util.Base64;
import java.util.List;

public class Utils {

	public static byte[] hexToBytes(String hex) {
		int len = hex.length();
		byte[] data = new byte[len / 2];
		for (int i = 0; i < len; i += 2) {
			data[i / 2] = (byte) ((Character.digit(hex.charAt(i), 16) << 4)
					+ Character.digit(hex.charAt(i+1), 16));
		}
		return data;
	}

	public static String bytesToHex(byte[] bytes) {
		StringBuilder result = new StringBuilder();
		for (byte b : bytes) {
			result.append(String.format("%02x", b));
		}
		return result.toString();
	}

	public static <T> T firstIntersection(List<T> list1, List<T> list2) {
		return list1.stream()
				.filter(list2::contains)
				.findFirst()
				.orElse(null);
	}

	public static byte[] getFileBytes(String filePath) throws IOException {
		try (FileInputStream fis = new FileInputStream(filePath)) {
			return fis.readAllBytes();
		}
	}

	public static InetSocketAddress parseSocketAddress(String socketAddress) {
		String[] split = socketAddress.split(":");
		String host = split[0];
		int port = Integer.parseInt(split[1]);
		return new InetSocketAddress(host, port);
	}
}
